AÑadido vista de tree a contact time y preferred degree
