+++ .settings/org.eclipse.core.resources.prefs
@@ -6,4 +6,6 @@ encoding//adm_uni/models/__init__.py=utf-8
 encoding//adm_uni/models/admission_application.py=utf-8
 encoding//adm_uni/models/admission_inquiry.py=utf-8
 encoding//adm_uni/models/attachment.py=utf-8
+encoding//adm_uni/models/contact_time.py=utf-8
+encoding//adm_uni/models/degree_program.py=utf-8
 encoding//adm_uni/models/languages.py=utf-8
