+++ adm_uni/models/__init__.py
@@ -2,6 +2,8 @@
 
 from . import languages
 from . import attachment
+from . import contact_time
+from . import degree_program
 from . import admission_inquiry
 from . import admission_application
 from . import res_partner
