+++ adm_uni/views/configuration.xml
@@ -20,5 +20,27 @@
             </tree>
         </field>
     </record>
+    
+    <record model="ir.ui.view" id="view_language_levels_tree">
+        <field name="name">Degree Program tree</field>
+        <field name="model">adm_uni.degree_program</field>
+        <field name="arch" type="xml">
+            <tree editable="bottom">
+                <field name="name" />
+            </tree>
+        </field>
+    </record>
+    
+    <record model="ir.ui.view" id="view_language_levels_tree">
+        <field name="name">Contact time tree</field>
+        <field name="model">adm_uni.contact_time</field>
+        <field name="arch" type="xml">
+            <tree editable="bottom">
+                <field name="name" />
+                <field name="from_time" />
+                <field name="to_time" />
+            </tree>
+        </field>
+    </record>
 
 </odoo>
